src/bbdeepMiRGene/requirements.txt
python=3.7
numpy
pandas
scikit-learn
keras
theano
biopython
viennarna
